Begin work on replacing accesses.
Simple read-only get() access works. Simple assignments with complex RHS
arguments also work, apart from pre-increment being handled incorrectly.
If there is a SINGLE pre-increment or pre-decrement operator on the RHS,
and no other references, the write is sequenced, and defined behaviour.
If there is more than one, or any post-increment/decrement, the write is
unsequenced. If there is a single pre-increment/decrement and another
read to the same field, it is also unsequenced and therefore undefined
behaviour. In case of undefined behaviour, unleash nasal demos.